^r Extract document/copy rendering and shared validation from render_injection_bundle.go, move-only (gofmt, vet, build, package unit+fuzz tests pass; declaration set verified identical; final package state byte-matches the verified full-split branch; internal restructure, no behavior change)
Step 3 of 3 for the injection god-file split, completing the iteration-5
god-file program: renderDocuments/renderCopies with container-target
validation move verbatim to render_documents_copies.go, and the shared
policy/path/secret validation helpers to render_validation.go, leaving the
entry points and manifest hashing in render_injection_bundle.go.
